refactor: move calculator functions from Controller to Statistics class
