+++ src/main/kotlin/controller/Controller.kt
@@ -1,7 +1,7 @@
 package controller
 
 import model.Lotto
-import model.Rank
+import model.Statistics
 import view.InputView
 import view.OutputView
 
@@ -36,40 +36,9 @@ class Controller {
         bonusNumber: Int,
         outputView: OutputView
     ) {
-        val matchResult = calculateMatchResults(lotto, winningNumbers, bonusNumber)
+        val matchResult = Statistics.calculateMatchResults(lotto, winningNumbers, bonusNumber)
         outputView.displayMatchResults(matchResult)
-        val winningStatistic = calculateWinningStatistic(lotto.purchaseAmount, matchResult)
+        val winningStatistic = Statistics.calculateWinningStatistic(lotto.purchaseAmount, matchResult)
         outputView.displayWinningStatistic(winningStatistic)
-    }
-
-
-    // TODO: move below helper functions out of Controller?
-    fun calculateMatchResults(lotto: Lotto, winningNumbers: List<Int>, bonusNumber: Int): List<Int> {
-        val matches = MutableList(6) { 0 }
-        for (ticket in lotto.tickets) {
-            when (Rank.valueOfEachTicket(ticket, winningNumbers, bonusNumber)) {
-                Rank.FIRST -> matches[0] += 1
-                Rank.SECOND -> matches[1] += 1
-                Rank.THIRD -> matches[2] += 1
-                Rank.FOURTH -> matches[3] += 1
-                Rank.FIFTH -> matches[4] += 1
-                Rank.MISS -> matches[5] += 1
-            }
-        }
-        return matches
-    }
-
-    fun calculateWinningAmount(matchResult: List<Int>): Int {
-        var totalAmount = 0
-        for ((index, rank) in Rank.entries.withIndex()) {
-            val count = matchResult.getOrElse(index) { 0 }
-            totalAmount += count * rank.winningMoney
-        }
-        return totalAmount
-    }
-
-    fun calculateWinningStatistic(purchaseAmount: Int, matchResult: List<Int>): Float {
-        val winningAmount = calculateWinningAmount(matchResult)
-        return (winningAmount / purchaseAmount).toFloat()
     }
 }
